+++ config/router.py
@@ -1,6 +1,6 @@
 from rest_framework import routers
 
-from auth.views import AuthViewSet
+from account.views import AuthViewSet
 from society.api.views import SocietyViewSet
 from student.api.views import StudentViewSet
 from society_manage.api.views import (
@@ -10,7 +10,7 @@ from society_manage.api.views import (
 
 router = routers.SimpleRouter()
 
-router.register(r'auth', AuthViewSet, base_name='auth')
+router.register(r'account', AuthViewSet, base_name='account')
 router.register(r'society', SocietyViewSet, base_name='society')
 router.register(r'student', StudentViewSet, base_name='student')
 router.register(r'society_manage/member', SocietyMemberViewSet, base_name='society_manage_members')
